mv utility methods into separate module
